requirements.txt
matplotlib==3.6.3
mediapipe==0.10.7
numpy==1.24.0
scipy==1.9.3
pandas==1.5.3
seaborn==0.11.2
umap-learn==0.5.7
mlflow==2.19.0
grad-cam==1.5.4
gdown==5.2.0
torch==2.2.2
torch-geometric==2.6.0
Flask==3.1.0
pillow==10.4.0
protobuf==3.20.3
azure-cosmos==4.9.0
dotenv==1.0.0

# torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2
# torch==2.0.1+cu118 torchvision==0.15.2+cu118 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118
